# internal/rest/openapi3.json
{"components":{"requestBodies":{"CreateUsersRequest":{"content":{"application/json":{"schema":{"properties":{"country":{"minLength":1,"type":"string"},"lastname":{"minLength":1,"type":"string"},"name":{"minLength":1,"type":"string"},"username":{"minLength":1,"type":"string"}}}}},"description":"Request used for creating a user.","required":true},"UpdateUsersRequest":{"content":{"application/json":{"schema":{"properties":{"country":{"minLength":1,"type":"string"},"lastname":{"minLength":1,"type":"string"},"name":{"minLength":1,"type":"string"},"username":{"minLength":1,"type":"string"}}}}},"description":"Request used for updating a user.","required":true}},"responses":{"CreateUsersResponse":{"content":{"application/json":{"schema":{"properties":{"user":{"$ref":"#/components/schemas/User"}}}}},"description":"Response returned back after creating users."},"ErrorResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"}}}}},"description":"Response when errors happen."},"ReadUsersByCountryResponse":{"content":{"application/json":{"schema":{"properties":{"user":{"$ref":"#/components/schemas/User"}}}}},"description":"Response returned back after searching users by country."},"ReadUsersResponse":{"content":{"application/json":{"schema":{"properties":{"user":{"$ref":"#/components/schemas/User"}}}}},"description":"Response returned back after searching one user."}},"schemas":{"User":{"properties":{"country":{"type":"string"},"id":{"format":"uuid","type":"string"},"lastname":{"type":"string"},"name":{"type":"string"},"username":{"type":"string"}},"type":"object"}}},"info":{"contact":{"url":"https://github.com/Oguzyildirim/go-crud"},"description":"REST APIs used for interacting with the User Service","license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"title":"User API","version":"0.0.0"},"openapi":"3.0.0","paths":{"/users":{"post":{"operationId":"CreateUser","requestBody":{"$ref":"#/components/requestBodies/CreateUsersRequest"},"responses":{"201":{"$ref":"#/components/responses/CreateUsersResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/users/by-country/{country}":{"get":{"operationId":"ReadUserByCountry","parameters":[{"in":"path","name":"country","required":true,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/ReadUsersByCountryResponse"},"404":{"description":"Users not found"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/users/{userId}":{"delete":{"operationId":"DeleteUser","parameters":[{"in":"path","name":"userId","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"description":"User updated"},"404":{"description":"User not found"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"get":{"operationId":"ReadUser","parameters":[{"in":"path","name":"userId","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/ReadUsersResponse"},"404":{"description":"User not found"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"put":{"operationId":"UpdateUser","parameters":[{"in":"path","name":"userId","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/UpdateUsersRequest"},"responses":{"200":{"description":"User updated"},"400":{"$ref":"#/components/responses/ErrorResponse"},"404":{"description":"User not found"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}}},"servers":[{"description":"Local development","url":"http://127.0.0.1:9234"}]}