# pkg/openapi3/user_types.gen.go
// Package openapi3 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.1 DO NOT EDIT.
package openapi3

// User defines model for User.
type User struct {
	Country  *string `json:"country,omitempty"`
	Id       *string `json:"id,omitempty"`
	Lastname *string `json:"lastname,omitempty"`
	Name     *string `json:"name,omitempty"`
	Username *string `json:"username,omitempty"`
}

// CreateUsersResponse defines model for CreateUsersResponse.
type CreateUsersResponse struct {
	User *User `json:"user,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error *string `json:"error,omitempty"`
}

// ReadUsersByCountryResponse defines model for ReadUsersByCountryResponse.
type ReadUsersByCountryResponse struct {
	User *User `json:"user,omitempty"`
}

// ReadUsersResponse defines model for ReadUsersResponse.
type ReadUsersResponse struct {
	User *User `json:"user,omitempty"`
}

// CreateUsersRequest defines model for CreateUsersRequest.
type CreateUsersRequest struct {
	Country  *string `json:"country,omitempty"`
	Lastname *string `json:"lastname,omitempty"`
	Name     *string `json:"name,omitempty"`
	Username *string `json:"username,omitempty"`
}

// UpdateUsersRequest defines model for UpdateUsersRequest.
type UpdateUsersRequest struct {
	Country  *string `json:"country,omitempty"`
	Lastname *string `json:"lastname,omitempty"`
	Name     *string `json:"name,omitempty"`
	Username *string `json:"username,omitempty"`
}

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody CreateUsersRequest

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody UpdateUsersRequest